refactor: extract scene editing feature boundary
