+++ src/pages/sceneEditor/sceneEditor.view.yaml
@@ -138,7 +138,7 @@ template:
                           - rtgl-text s=sm: ...
                       - rtgl-view w=1fg: null
                   - rtgl-view w=f p=md sv h=1fg:
-                      - rvn-lines-editor#linesEditor :lines=currentLines :selectedLineId=selectedLineId :sectionLineChanges=sectionLineChanges :repositoryState=repositoryState: null
+                      - rvn-lines-editor#linesEditor :lines=currentLines :selectedLineId=selectedLineId: null
                       - rtgl-view h=30vh: null
           - rtgl-view w=1 h=f bgc=mu: null
           - rtgl-view w=f:
